requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
google-api-python-client==2.121.0
google-auth==2.29.0
bs4